Split the card per component and merge into intem
